Add AppIconTitle property wrapper. Remove AppSettings.
