+++ Nio/Conversations/RecentRoomsView.swift
@@ -6,7 +6,6 @@ import NioKit
 
 struct RecentRoomsContainerView: View {
     @EnvironmentObject var store: AccountStore
-    @EnvironmentObject var settings: AppSettings
     @AppStorage("accentColor") var accentColor: Color = .purple
 
     @State private var selectedNavigationItem: SelectedNavigationItem?
@@ -18,7 +17,6 @@ struct RecentRoomsContainerView: View {
                 NavigationSheet(selectedItem: $0)
                     // This really shouldn't be necessary. SwiftUI bug?
                     .environmentObject(self.store)
-                    .environmentObject(self.settings)
                     .accentColor(accentColor)
             }
             .onAppear {
